Make closure compiler a variable in the Makefile.
This way, users can pass different things as the command to invoke the
closure compiler.  For example, "java -jar /path/to/compiler.jar" or
something like that.

Introduced CLOSURE_ARGS parameter so users can pass additional arguments to
the closure compiler, e.g.
CLOSURE_ARGS="-O=WHITESPACE_ONLY --formatting=PRETTY_PRINT"
to get a human-readable version of the almost unmodified input for debugging.

Got rid of unneccessary perl invocation.  We don't need --js in front of
every single file; once for the whole set is enough.  (Even if we wanted it
for every file, we still could do so without perl).

Dropped the double colon from the rule.  There appears to be no need for
such an obscure feature here.  The other double-colon rules are still in
place, though.